feat: audit tests optional

+++ tests.sh
@@ -2,8 +2,10 @@
 
 set -e
 
+RUN_AUDIT_TESTS=${RUN_AUDIT_TESTS:-1}
+
 START_TIME=$(date +'%Y-%m-%d %H:%M')
-echo "Starting tests at $START_TIME" | tee -a ./logs/summary.log
+echo "Starting tests at $START_TIME" | ts '[%Y-%m-%d %H:%M:%S]' | tee -a ./logs/summary.log
 
 LABEL_APP_KEY=$(kubectl get ds falco -n falco -oyaml | yq '.metadata.labels' | grep -v 'instance' | grep 'falco' | grep 'app' | yq '. | keys' | cut -c 3-)
 echo "Using label key for selector: $LABEL_APP_KEY"
@@ -47,35 +49,39 @@ else
   echo "[ FAIL ]: Detect attach/exec to pod" | ts '[%Y-%m-%d %H:%M:%S]' | tee -a ./logs/summary.log
 fi
 
-echo "Checking Falco service"
-TEST_SVC=$(kubectl get svc falco -n falco 2>/dev/null ||:)
-if [ "$TEST_SVC" != "" ]; then
-  echo "[  OK  ]: Falco service deployed" | ts '[%Y-%m-%d %H:%M:%S]' | tee -a ./logs/summary.log
-else
-  echo "[ FAIL ]: Falco service not deployed" | ts '[%Y-%m-%d %H:%M:%S]' | tee -a ./logs/summary.log
-fi
+if [ $RUN_AUDIT_TESTS -ne 0 ]; then
 
-echo "Cleaning up possible debug pod"
-kubectl delete pod debug -n kube-system 2>/dev/null ||:
+  echo "Checking Falco service"
+  TEST_SVC=$(kubectl get svc falco -n falco 2>/dev/null ||:)
+  if [ "$TEST_SVC" != "" ]; then
+    echo "[  OK  ]: Falco service deployed" | ts '[%Y-%m-%d %H:%M:%S]' | tee -a ./logs/summary.log
+  else
+    echo "[ FAIL ]: Falco service not deployed" | ts '[%Y-%m-%d %H:%M:%S]' | tee -a ./logs/summary.log
+  fi
 
-echo "Launching pod in kube-system"
-TEST_START=$(date +'%Y-%m-%dT%H:%M:%SZ' --utc)
-kubectl run -it --rm --restart=Never debug -n kube-system --image alpine -- ls 1>/dev/null
-echo -n "Checking kube audit detection"
-MATCH="Warning Pod created in kube namespace (user=system:admin pod=debug ns=kube-system images=alpine)"
-TEST_AUDIT=$(kubectl logs daemonset/falco -n falco --since-time="$TEST_START" | tee logs/detect_audit.log | grep "$MATCH" ||:)
-I=20
-while [ $I -ne 0 ] && [ "$TEST_AUDIT" == "" ]; do
-  sleep 3
+  echo "Cleaning up possible debug pod"
+  kubectl delete pod debug -n kube-system 2>/dev/null ||:
+
+  echo "Launching pod in kube-system"
+  TEST_START=$(date +'%Y-%m-%dT%H:%M:%SZ' --utc)
+  kubectl run -it --rm --restart=Never debug -n kube-system --image alpine -- ls 1>/dev/null
+  echo -n "Checking kube audit detection"
+  MATCH="Warning Pod created in kube namespace (user=system:admin pod=debug ns=kube-system images=alpine)"
   TEST_AUDIT=$(kubectl logs daemonset/falco -n falco --since-time="$TEST_START" | tee logs/detect_audit.log | grep "$MATCH" ||:)
-  let I=I-1
-  echo -n "."
-done
-echo ""
-if [ "$TEST_AUDIT" != "" ]; then
-  echo "[  OK  ]: Detect pod created in kube namespace" | ts '[%Y-%m-%d %H:%M:%S]' | tee -a ./logs/summary.log
-else
-  echo "[ FAIL ]: Detect pod created in kube namespace" | ts '[%Y-%m-%d %H:%M:%S]' | tee -a ./logs/summary.log
+  I=20
+  while [ $I -ne 0 ] && [ "$TEST_AUDIT" == "" ]; do
+    sleep 3
+    TEST_AUDIT=$(kubectl logs daemonset/falco -n falco --since-time="$TEST_START" | tee logs/detect_audit.log | grep "$MATCH" ||:)
+    let I=I-1
+    echo -n "."
+  done
+  echo ""
+  if [ "$TEST_AUDIT" != "" ]; then
+    echo "[  OK  ]: Detect pod created in kube namespace" | ts '[%Y-%m-%d %H:%M:%S]' | tee -a ./logs/summary.log
+  else
+    echo "[ FAIL ]: Detect pod created in kube namespace" | ts '[%Y-%m-%d %H:%M:%S]' | tee -a ./logs/summary.log
+  fi
+
 fi
 
 echo "Tests finished"
